Fix lint and E2E test failures for template-backed quick responses
Extract TemplateCard component from TemplatesPanel to resolve max-lines
lint error, fix extra line break in NewSessionView, and update
quick-response-focus-behavior E2E tests to use template-backed seeding
instead of standalone quick responses.
