+++ tests/e2e/quick-response-focus-behavior.spec.ts
@@ -2,7 +2,7 @@ import { test, expect } from '@playwright/test';
 import {
   seedProject,
   seedSession,
-  seedQuickResponse,
+  seedProjectTemplate,
   cleanupCreatedResources,
   waitForPageReady,
   openSessionOverlay,
@@ -15,7 +15,29 @@ import {
  * - Non-auto-submit responses blur the textarea (no mobile keyboard)
  * - Auto-submit responses still submit the form correctly
  * - Content is inserted correctly in both cases
+ *
+ * Uses template-backed quick responses (showInQuickResponses: true).
  */
+
+const createdTemplateIds: string[] = [];
+
+async function seedQuickResponseTemplate(
+  projectId: string,
+  data: {
+    label: string;
+    content: string;
+    autoSubmit?: boolean;
+  }
+) {
+  const template = await seedProjectTemplate(projectId, {
+    name: data.label,
+    prompt: data.content,
+    showInQuickResponses: true,
+    quickResponseAutoSubmit: data.autoSubmit ?? false,
+  });
+  createdTemplateIds.push(template.id);
+  return template;
+}
 
 test.beforeEach(async () => {
   await cleanupCreatedResources();
@@ -28,13 +50,20 @@ test.afterEach(async () => {
 test.describe('Quick Response Focus Behavior', () => {
   test('non-auto-submit response blurs textarea in new session view', async ({ page }) => {
     const project = await seedProject('QR Focus NewSession', '/tmp/qr-focus-new');
-    await seedQuickResponse(project.id, {
+    await seedQuickResponseTemplate(project.id, {
       label: 'No Auto',
       content: 'Content without auto-submit',
       autoSubmit: false,
     });
 
+    // Wait for templates API before interacting
+    const apiDone = page.waitForResponse(
+      (resp) => resp.url().includes('/templates') && resp.status() === 200,
+      { timeout: 30000 }
+    );
+
     await page.goto(`/projects/${project.id}/sessions/new`);
+    await apiDone;
     await waitForPageReady(page);
 
     // Expand quick responses panel
@@ -60,7 +89,7 @@ test.describe('Quick Response Focus Behavior', () => {
 
   test('non-auto-submit response blurs textarea in conversation view', async ({ page }) => {
     const project = await seedProject('QR Focus Conversation', '/tmp/qr-focus-conv');
-    await seedQuickResponse(project.id, {
+    await seedQuickResponseTemplate(project.id, {
       label: 'No Auto Conv',
       content: 'Conversation content',
       autoSubmit: false,
@@ -71,16 +100,16 @@ test.describe('Quick Response Focus Behavior', () => {
       startImmediately: false,
     });
 
-    // Set up the quick-responses API response listener BEFORE navigating
+    // Set up the templates API response listener BEFORE navigating
     const apiDone = page.waitForResponse(
-      (resp) => resp.url().includes('/quick-responses') && resp.status() === 200,
+      (resp) => resp.url().includes('/templates') && resp.status() === 200,
       { timeout: 30000 }
     );
 
     await page.goto(`/sessions/${session.id}/summary`);
     await openSessionOverlay(page);
 
-    // Wait for the quick-responses API call to complete
+    // Wait for the templates API call to complete
     await apiDone;
 
     // Wait for the panel to render
@@ -106,13 +135,20 @@ test.describe('Quick Response Focus Behavior', () => {
 
   test('auto-submit response still submits form in new session view', async ({ page }) => {
     const project = await seedProject('QR Focus AutoSubmit', '/tmp/qr-focus-auto');
-    await seedQuickResponse(project.id, {
+    await seedQuickResponseTemplate(project.id, {
       label: 'Auto Submit',
       content: 'Auto-submit content',
       autoSubmit: true,
     });
 
+    // Wait for templates API before interacting
+    const apiDone = page.waitForResponse(
+      (resp) => resp.url().includes('/templates') && resp.status() === 200,
+      { timeout: 30000 }
+    );
+
     await page.goto(`/projects/${project.id}/sessions/new`);
+    await apiDone;
     await waitForPageReady(page);
 
     // Expand quick responses panel
@@ -131,7 +167,7 @@ test.describe('Quick Response Focus Behavior', () => {
 
   test('mobile viewport - non-auto-submit blurs textarea', async ({ page }) => {
     const project = await seedProject('QR Focus Mobile', '/tmp/qr-focus-mobile');
-    await seedQuickResponse(project.id, {
+    await seedQuickResponseTemplate(project.id, {
       label: 'Mobile Test',
       content: 'Mobile content',
       autoSubmit: false,
@@ -140,7 +176,14 @@ test.describe('Quick Response Focus Behavior', () => {
     // Set mobile viewport
     await page.setViewportSize({ width: 375, height: 667 });
 
+    // Wait for templates API before interacting
+    const apiDone = page.waitForResponse(
+      (resp) => resp.url().includes('/templates') && resp.status() === 200,
+      { timeout: 30000 }
+    );
+
     await page.goto(`/projects/${project.id}/sessions/new`);
+    await apiDone;
     await waitForPageReady(page);
 
     // Expand quick responses panel
